Instalog: add two-level archive path in report event
The processed event contains the gcs path for downloading the archive
and where the report resides in the archive. For two-level archive, we
need to record the full path in parent archive. Extend the
`ReportParser` to have an extra field for recording the full path in
first level archive, and append it to the final report event.

BUG=b:225278303
TEST=make test

Change-Id: I9781690b129bdb1aac8b40de3a4dfc88eb0d0254
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3743831
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>