gooftool: Add skip_enable_check for testing
Add an flag and parameters to crosfw that can skip the final check,
which expects a failure to disable the write protect after enabling the
write protect on target. This feature is only used for CI testing, as we
will perform enable and disable software write protect back-and-forth
during the test while the hard write protect is turned off at then.

BUG=b:231274641
TEST=(1) Run make test.
     (2) While HWWP is on, run `gooftool write_protect` on DUT can
         enable the write protect on all targets.
     (3) While HWWP is off, run `gooftool write_protect --targets AP
         --skip_enable_check` does not trigger the check, and the write
         protect status of AP is in the expected on status.

Change-Id: Ic852c95fc33fac505efaa71829137d2437d6f8e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3754991
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>