gooftool: Implement getting the WP status in different target
Gooftool is to be extended as an adapter for the command `flashrom`,
which can read the write protection status of different targets.
Implement getting the write protect status on AP, EC, PD, and FPMCU in
the `WriteProtectTarget` to be used by Gooftool. For target not on the
DUT, or not yet implemented the method for getting the detail of the
write protect status, an `UnsupportedOperationError`
exception is raised.

BUG=b:231274641
TEST=make test

Change-Id: Ib276b7e38e29821df880956e25775dd25948fe9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3754989
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>