update_cr50_firmware: set the default ti50 firmware path
In this CL, we set the default ti50 firmware path to
`/opt/google/ti50/firmware/ti50.bin.prod`. If `is_ti50` is set to
true and `firmware` is not set, then the test will use the default ti50
firmware to update the ti50. The test will also check if the update
is valid. (e.g. to update to ti50 RW 0.0.16+, user must update to RW
0.0.15 first.)

BUG=b:236793753
TEST=run on nirwen

Change-Id: I9bea20adac6b396ff281e0ea38baa940acbcdd95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3758445
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>