installer: specify the target path for extracted makeself content
The `/tmp` is restricted with following the symbolic link, which breaks
the installation process to load the installer.py to do the job. Specify
the `target` option to `/usr/local/tmp` while building the toolkit
installer so that symbolic link works during the installation.

BUG=b:240875191
TEST=make toolkit and upload to DUT, can install the toolkit.

Change-Id: Ib702ec5f8411e62311f8b4ca6f404a8bd5febb99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3800569
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>