hwid/v3: use hash value to get up-to-date comp names
In change unit splitting process, the component name and sequence number
of a component is non-deterministic.  In order to have an identifier to
locate components at runtime, class Components provides a method
GetComponentNameByHash which uses the hash of dumped component as the
unique identifier to return the component name of such component.

BUG=b:232063010
TEST=make test

Change-Id: Ibfbceef2697a7363e042978eefcc17d22f818185
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3819449
Tested-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>