installer: update extracted path and add cleanup procedure.
Makeself first extracts contents to a path, then starts the
installation process. It does not overwrite the existing content if some
files are under the specified path, thus extra files remain and are
installed to the system unexpectedly. Update the path for extract
content for easier cleanup, add the cleanup script, and specify the
`--cleanup` option while building the installer.

BUG=b:240875191
TEST=(1) Perform installation and ensure the extracted contents are
         removed.
     (2) Ensure `factory_toolkit.run --lsm` outputs correct info.

Change-Id: Ia3eca1c23fee639b918511d396c933d32f9b5491
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3823538
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>