isort: Print a warning instead of blocking presubmit when fail
We didn't find a good approach to solve the conflict between isort and
yapf. In the case that a file pass by yapf but not isort, we mark the
file uncertain and warn people to check manually instead of blocking
presubmit.
To keep the feature of isort, isort will still be applied when we use
make format to fix codes, but it will be overridden by yapf if there's
a conflict between isort and yapf.
Since we can handle the conflict, revert the change that modify isort
config to try to make isort compatible with yapf.

BUG=b:226500333
TEST=Check message when files pass yapf but not isort
TEST=Check message when files not pass yapf and isort
TEST=The special case in the issue can pass presubmit with warning

Change-Id: If4726128cebc953e468fc0de0e962362862ffcda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3823547
Commit-Queue: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>
Auto-Submit: Chih-Yao Chuang <jasonchuang@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>