wipe: raise WipeError when to_stop_service_list is not empty
Currently, we do not check if `to_stop_service_list` is empty, and
services might still be alive after we try to stop it. We should fail
immediately if the list is not empty, and this will also make debugging
easier.
We also add the job to instance key mapping, since we cannot get the
instance key from `initctl`.

BUG=b:239799995
TEST=(1) `start ml-service TASK=adaptive_charging`
         `gooftool wipe_in_place --test_umount` fails
     (2) `stop ml-service TASK=adaptive_charging`
         `gooftool wipe_in_place --test_umount` succeeds

Change-Id: I5d7e0ee06985000894a9c78ea7724ad82dd0ecdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3829321
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Tested-by: Phoebe Wang <phoebewang@chromium.org>