hwid/v3: add dependency relations to change units
This CL introduces the DependencyNode class and uses topological sort to
collect as many as auto-mergeable change units.  These change units can
be used to create a CL without reviews.   The remaining change units
would be used to create another CL based on the previous one for
reviewers to focus on the important ones.

BUG=b:232063010
TEST=make test

Change-Id: Id0e9c44ecd3818d2c3088621c277f6b23175f581
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3856280
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>