default_probe_statement: Add usb device information for ethernet
As the internal usb ethernet controller allowed to be probed, we need to
include the information of usb devices (e.g. "idProduct", "idVendor",
and "bcdDevice") into the result of the probe function of ethernet.

BUG=b:242006801
TEST=probe ethernet on DUT and make sure idProduct, idVendor, and
bcdDevice are shown in the result

Change-Id: I40f3797cd457a732ab00b1c13c58d95af771da9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3869029
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Jeff Lin <jeffulin@google.com>
Auto-Submit: Jeff Lin <jeffulin@google.com>
Commit-Queue: Jeff Lin <jeffulin@google.com>