Instalog: input_gcs_scan: Check the stability of new files
If partner is uploading an archive to SFTP, Google Cloud Storage may
show a corrupted archive. The report parser scan the GCS every 60
seconds, so it may download and parse incompleted archives.
To avoid this issue, the plugin check the stability of new files before
it emits new files to other plugins. Also extend the default interval to
600 seconds.
If a file is changed on GCS, the plugin will re-emits it.

BUG=b:247944228
TEST=manually test

Change-Id: I0b1295d9279a18f0f7f0585749279f58e150d19e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3906799
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Commit-Queue: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>