gooftool: Cleanup gooftool write_protect_target
This CL is a further code cleanup after the refactoring of gooftol
write_protect_target. This CL deals with typo, comment, and syntax
related issues only, and does nothing with logic or workflow.

This CL udpates the header comments for write_protect_target so that
they meet the latest rule.

This CL adds a missing prefix underscore for each WriteProtectTarget
implementation class which is not exported to external scripts.

This CL adds more typings to function parameters and return value.

BUG=b:239791358
TEST=gooftool write_protect --targets FPMCU -v 4

Change-Id: Idee6f9506318ccde1bc91d98c48619943b54afab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3916000
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>