gooftool: device-tree-compatible-match is now frid
On ToT, device-tree-compatible-match is renamed to frid. For those
version elder than 15105.0.0 start to include new config.yaml in the
release image.

In gooftool, we compares the identity in config.yaml of release image
and config.yaml of test image. If they don't match, the test will fail.
Therefore, we need to "normalize" the attribute name between release
image and test image, so they can match even though one is
`device-tree-compatible-match`, and the other one is `frid`.

BUG=b:250470707
TEST=None

Change-Id: I82926d36e3a6cce6d7191935efc70412a4b72170
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3932846
Commit-Queue: Bob Moragues <moragues@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Tested-by: Kenny Pan <kennypan@google.com>