hwid/service: Fix battery payload generator with truncated values
When recording values to HWID DB, the trailing spaces will be trimmed,
and those values will be considered as short values incorrectly.  For
example, assuming the value in EC is "ABC-56 8", then the value from
sysfs will be "ABC-56 " and "ABC-56" will be stored to the HWID DB.

The generator will fill spaces and do a prefix match when values are
shorter than truncated length.

BUG=b:249698819
TEST=unit test

Change-Id: I13d44c2db40d6bcf7020f60c5a5e5bef60ada915
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3935953
Tested-by: Kevin Lin <kevinptt@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Kevin Lin <kevinptt@chromium.org>