instalog: fix exception using external tool checking empty zip archive
Instalog cross-checks number of factory reports in the downloaded zip
with external tool `unzip`. If the zip is empty, calling `unzip` gets
process exist status 1, yet this is not an error. Exclude the such case
when calling `unzip` by checking whether the exit status is 1, and the
string `zipfile is empty` is in stderr.

BUG=b:251744419
TEST=make test

Change-Id: Ie007db144478ccb0011a3f7ae594993bc8dacfa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3944768
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>