hwid/service: handle patches to initial DB specially
The first change after initial DB behaves differently compared to later
CLs.  For example, the order and bit length in bit patterns should not
be reordered/resized.  This CL uses the emptiness of PatternDatum.fields
as a flag indicating that the DB change is applied to an initial DB and
preserves the bit pattern uploaded by users.

BUG=b:252919734
TEST=make test
TEST=deploy/cros_hwid_service.sh test

Change-Id: Ia0e8fc588d47d77ac036b571e0a76a1d2e729c8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3948422
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>