model_sku_utils: Match on the design name as well for product_sku.
When matching on FRID instead of smbios name or device tree compatible
match, the smbios name and device tree compatible match values are not
available to use in the factory. Instead, with CL:3971464, the design
name will be used for FRID-matching designs when generating the
product_sku keys, just like model and oem_name.

BUG=b:253966060
TEST=make test

Change-Id: I96df5ca4bf35ccdeec1690cd939c72698be2c214
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3974310
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>