pytests: Make cursor invisible in display test
In display tests, users don't want to see the cursor
when a full screen image is displayed because that
the cursor may block the bad pixels on the display.
This CL modifies "cursor" attribute to make the cursor
invisible in display and display_point tests.

BUG=b:258374239
TEST=(1) make test
     (2) Manually test on pujjoteen DUT.

Change-Id: I941457fb34548b881537e3da23fddb945ca48d20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4022993
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Stimim Chen <stimim@chromium.org>