py: Allow factory test to waive itself in pytest
Currently we allow tests to be "waived" by setting "waived_tests"
and "conditional_patches" params in "test_list".
However, when a test item is 'waived', it'll pass for no matter reasons,
including syntax error or segv in execution in pytest.
Besides, there may be cases "we want to waive particular item or test".
This CL adds a "WaiveTask" function in "TestCase", which allows the
factory test to waive itself in pytest.

BUG=b:172200663
TEST=(1)make test
     (2)manually test on pujjoteen DUT

Change-Id: Ib4c63e07c82e024e20c1c235b5dba48781473379
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4046383
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Stimim Chen <stimim@chromium.org>