tools: Save factory toolkit in factory_bug
Currently we collect logs by running 'factory_bug' command.
Now we want to improve the issue template to provide more
sufficient information for debugging, and one of the extra
information we want is the factory toolkit.

This CL adds 'SaveFactoryToolkit' functions for packing
the code of factory toolkit on device as 'factory_toolkit.zip'.
It let the ODMs to provide more information easier.

BUG=b:260057902
TEST=(1)make test
     (2)manually test on pujjoteen DUT

Change-Id: Idd2bda8d344be681b6313604797b46420504dd4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4054430
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>