test utils: csv_utils to async record and upload CSV entries
To upload a CSV entry via `umpire_server_proxy.UploadCSVEntry`,
currently the pytest needs to connect to the network for the function to
work.  But in a real factory, devices are only connected to the network
at certain stations when necessary. In this case, partners would be
reluctant to run tests that requires network access in the middle of
FATP, or they would try to modify the test locally to skip CSV
uploading.

In this CL, we introduce a new way to upload CSV entries in two steps:
1. `csv_utils.CSVManager.Append` to add an entry to a CSV file.
2. `csv_utils.CSVManager.UploadAll` to upload all saved entries to
   factory server. The uploaded entries will be deleted from local file.

`csv_utils.CSVManager.UploadAll` is added to the `sync_factory_server`
pytest.

BUG=None
TEST=make test
TEST=manual test on device: write a dummy pytest to call the save
     function, and then run the sync_factory_server test.
TEST=for i in $(seq 10); do
       PYTHONWARNINGS="error::ResourceWarning" bin/factory_env \
           py/test/utils/csv_utils_e2e_test.py || break
     done

Change-Id: I087260d9c3d614dca49ad9ee56d2de66969dee53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4132204
Auto-Submit: Stimim Chen <stimim@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>