hwid/v3: disallow component name and values to be modified together
When users mistakenly update a HWID database contents based on a legacy
snapshot, it often generates incompatible changes.  One common pattern
is that the user inserts a new component to the database instead
of appending at the end.  The HWID framework considers those kind of
change as modifying the component name and values at the same time.

As it's uncommon for a regular change proposal to modify both the
component name and the values, this CL simply rejects the scenario
to mitigate the chance to get incompatible HWID database changes.

BUG=b:261142651
TEST=make test

Change-Id: I2374879811a2ebf4ff6e6f4f175fe67b9dff4c6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4088247
Reviewed-by: Yu-An Wang <wyuang@google.com>
Tested-by: Yong Hong <yhong@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>