image_tool: verify cros config behavior enhancements
Currently image tool is pre-checking too strictly on cros_configs
settings. The config acquired from test image has to be the exactly
the same as the config acquired from FSI in nowadays pre-check. We
reduce the level of the pre-check by only fobidding the cases
where one of the config in test image not found in FSI. This
guaranteed that whenever this bundle passes factory process,
we will be confident that at least the SKU would be matched in FSI.

Also, as introduced in b:245588383, smbios-name-match and
device-tree-compatible-match is now frid. It is painful for
partners unabling to create bundle with old factory images +
FSI using frid and this is quite common. We perform a key
transform here to make this kind of bundle creation possible.

BUG=b:258155546
TEST=Tested with total 6 cases
smbios/dt test image + smbios/dt FSI
frid test image + frid FSI
smbios/dt test image + frid FSI
for the above 3 cases we tested on both x86 and arm
expected result: being able to create image

Change-Id: Id4b7171d2e189c9432a579dd6f4dd5ca0423d951
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4087704
Tested-by: Wei-Shuo Su <stevesu@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>