gooftool: modify gooftool verify_cros_config behavior
Frid was introduced in 15227.0.0 while smbios-name-match /
device-tree-compatible-match was removed. As a result, it is
expected that the original exact match way of verifying
cros config would not work.

As we look into this failure, we came to realize that our way
of verification is different from how crosid match identities.
In this CL, we are changing our verification in a way that resemble
to how crosid matches identities. We now regard identity collected
from DUT as source of truth and whenever we could match something
in test image config.yaml and FSI config.yaml while also their
brand-code is the same, we should let it pass, as in this case
we're assuered that it will match config when boot into	FSI.

BUG=b:295267911
TEST=(1) Install factory branch test image
     (2) Check that `crosid` match something
     (3) Install release image with correct identity
         via check_image_version.py
     (4) See `gooftool verify_cros_config` pass
     (5) Install release image without correct identity
         via check_image_version.py
     (6) See `gooftool verify_cros_config` fail

Change-Id: Ie544e33ac66a16298e3b5cf359d0e99b17dbe621
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4768976
Auto-Submit: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>