[WORKAROUND] init: minijail added a new argument
In https://crrev.com/c/4024902, a new argument
'no-default-runtime-environment' was added to minijail.
Then, in https://crrev.com/c/4043304, powerd added this argument to its
upstart conf file.

In factory, we wrap minijail to "jailbreak". However, the wrapper
doesn't recognize the new argument.  Therefore, pwoerd is not running
when factory toolkit is enabled.  The first side effect we found is
b/261790474.  The chrome browser relies on powerd to get the status of
"tablet mode"

BUG=b:261790474
TEST=tested on device

Change-Id: I7d2b1a129a5a75158a4485a3a01739adb39cd154
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4091104
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tzu-Min Sun <jimmysun@chromium.org>