goofy: Stop running tests before clearing status
The original behavior of clearing status is confusing when the user
clear the status of a running test. The selected test will be stopped
and cleared, but goofy will continually activate the next tests if it's
already enqueued. This CL make goofy stop all the other tests before
clearing status.

BUG=b:212503815
TEST=(1) make test
     (2) Manually try to clear the status of a running tests

Change-Id: I065d3bc798c4a93c1465624af85c15e6f220c6fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4096160
Commit-Queue: Jeff Lin <jeffulin@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Jeff Lin <jeffulin@google.com>