goofy: change connection_manager behavior
wipe_in_place wants to disable modemmanager while keep network
connected. Meanwhile, connection_manager will re-enable modemmanager on
stop that results in a dependency conflict.

The goal of enabling network on stop of connection_manager is to have
the connection until the last minutes so it's easier to debug.

Thus, this CL changes the behavior by:
1. makes the connection_manager only disable network on stop
when Goofy is running.
2. adds an additional condition that only re-enable network when the
network manager is not running due to some unexpected reason as this
should be the only situation we want to re-enable the network.

BUG=b:262931357
TEST=(1) gooftool wipe_in_place --test_umount
     (2) initctl list | grep modemmanager
     (3) check modemmanager is stopped
TEST=(1) stop modemmanager
     (2) stop factory
     (3) initctl list | grep modemmanager
     (4) check modemmanager is stopped

Change-Id: Ia6326cdc2ece76e4df0d69f10a9d6742ac2fe64b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4113046
Commit-Queue: Yu-An Wang <wyuang@google.com>
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Tested-by: Yu-An Wang <wyuang@google.com>