mrc_cache: cache the last line of eventlog
Current implementation relies on the order of `System boot` and
`MRC Cache Update` event. This is not robust since the order of the
event log might change.
This CL caches the last line of event log before reboot and compares it
against the newly generated logs.

BUG=b:264971516
TEST=run on DUT

Change-Id: I085350d7a1f70c01b7841a935ab00969eec62eac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4151675
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: TH Lin <t.h_lin@quanta.corp-partner.google.com>
Tested-by: TH Lin <t.h_lin@quanta.corp-partner.google.com>