record: replace the `time` field of event type `station.test_run`
The `time` field of a record should store the timestamp that the event
is generated. However, the `time` field of the `station.test_run` event
stores the time that a factory test starts, rather than the time that a
test ends. This CL replaces field `time` with `endTime` in this case.

Below are two `station.test_run` events. From the example we can see
that their timestamps are the same:
```
// Test starts
{"status": "STARTING", "time": 1675736066.5407052}
// Test ends
{"status": "PASS", "time": 1675736066.5407052, "endTime":
1675736066.7728324}
```

BUG=b:264354294
TEST=make test

Change-Id: I2c7b065b5f4442964ee501407a1b6fadf0233c71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4229945
Reviewed-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>