touchpad: fix ungrab not claimed evdev device
The touchpad test grabs the input during the test run, and ungrabs it
in `tearDown` test hook. However, if the test is interrupted before
claiming the device input, the `ungrab` call in `tearDown` still runs
and raises exception, thus skips the following clean up process, such
as restoring the lid open, etc.

This commit replaces the claim of the resource with context manager
fashion, such that the grab and ungrab is handled by the context
manager, and only ungrab the device once grabbed.

BUG=b:253948622
TEST=run touchpad in toolkit with
     (1) follow the test test procedure on screen, and can complete the
	 test without getting any exception in `tearDown` hook.
     (2) click mark as failed before claiming the device, and no
         exception is raised in `tearDown` hook.

Change-Id: I78b5c3c4adb628cc98467be5669634c378a05565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4235677
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Wei-Shuo Su <stevesu@google.com>