hwid/service: Update battery payload generator to skip technology field
The fields technology and chemistry are fixed values in general, which
means we don't need to generate probe statements for these two fields.
This CL modifies the payload generator to set the chemistry field
optional if the value in HWID DB is not vendor-specific, otherwise still
generate probe statements for it.

BUG=b:265395014
TEST=./deploy/cros_hwid_service.sh test
TEST=make test

Change-Id: I557d081c9d20361d2f883121ef3dafc4b1025d38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4236319
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>