finalize_bundle: add new script to extract fw info
The arguments of --extract-firmware-info are different from the
oringinal usage of finalize_bundle. For example, MANIFEST should not be
required to extract firmware info and most arguments are not effective
for --extract-firmware-info.

This CL splits the usage to a new script extract_firmware_info to make
the argument clearer.

BUG=b:269971864
TEST=bin/extract_firmware_info

Change-Id: I16fa30fcab4c6205eaea35cbe27bccc2bb524595
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4258347
Commit-Queue: Yu-An Wang <wyuang@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Yu-An Wang <wyuang@google.com>