py: Update GetSystemDetails() with factory_summary
Extracts the code of getting system_info to a function GetSystemInfo().
Change the type of BLOCKLIST_KEYS from list to set for better time
complexity in the future extension.
Finally adds the output of factory_summary to GetSystemDetails(), and
updates the unittest of it.

BUG=b:269966106
TEST=(1)make test.
     (2)manually test on steelix DUT and make sure it work as expected.

Change-Id: I12362450b8d0112b3c518b0ae820440eec5a5d04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4274861
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>