sync_factory_server: Fix change url behavior
In CL:3948742, we've modified the behavior of exception handling in `sync_utils.EventWait`. This cause the code to raise exception when retry timeouts, which will cause the `SyncFactoryServer` task to fail immediately. This CL fixes the behavior by properly handle the `TimeoutError` error and the return value when `event` is not set.

BUG=b:271317349, b:222638799
TEST=make test
TEST=Run the change on DUT and verify the URL can be set.

Change-Id: Ic11d03cdad8a2a04b7ac55bde3b68ab5f56cba68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4296862
Commit-Queue: Louis Chiu <louischiu@google.com>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Tested-by: Louis Chiu <louischiu@google.com>