process_utils: Add typings and fix type-related bugs
This CL adds typings for `process_utils`. Some bugs concerning types are
fixed along with the added typings.

The old behavior of `process_utils.Spawn` expects that type of the
argument `args`, which is the command to be executed, to be a string
when `shell` is set or else a list of string. But the processing logic
does not handle this well so that when `shell` is not set `args` can
also be a string. This CL adds strictly type checking for `args` and
fixes misusing callers.

This CL updates docstring of `process_utils.Spawn` so that comment
symbol `#` is not interpreted as markdown header syntax for IDE.

This CL exposes `_ExtendedPopen` and renames it to `ExtendedPopen` since
many callers are using it.

BUG=b:273881230
TEST=make test

Change-Id: I08255f39d54edb43a2e3757e4ee1d355a4bf57d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4337297
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Louis Chiu <louischiu@google.com>
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>