py: Add GetSystemInfo() in common.Util
'cros.factory.tools.factory_summary' was imported into 'core.py'
in CL:4274861, which causes the fail of dependency checking
in presubmit.
According to README.md of gooftool, gooftool should not use
Factory Test or Device API.

So we make the following changes for it:
1. Replace the word 'info' by 'summary' in the method name
   in factory_summary for style consistency.
2. Create a method called GetSystemInfo() in common.Util,
3. Move the code snippets from factory_summary.GetSystemSummary()
   to common.Util.GetSystemInfo().
4. Call common.Util.GetSystemInfo() in
   factory_summary.GetSystemSummary() and
   core.Gooftool.GetSystemDetails().

We also integrate the additional system information in
core.Gooftool.GetSystemDetails() to common.Util.GetSystemInfo(),
which is one of the tasks in the issue.

BUG=b:269967293
TEST=(1)make test
     (2)Manually run 'factory_summary system' on corsola DUT.

Change-Id: I31818014cb35134e8beac7a06e7bd062babf2eb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4325367
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>