bundle_creator: add cc_emails in create bundle request
Add `cc_emails` field in both create bundle requests.  The field is
used while sending a result mail, it supports sending the carbon copy
to the emails listed in `cc_emails`.

In this change it also moves the failure email from `to` to `cc`.
And because of that `cc` couldn't be an empty list while calling
`send_mail`, use `kwargs` instead.

BUG=b:264512542
TEST=(1) ./deploy/bundle_creator.sh deploy-appengine-legacy dev
     (2) ./deploy/bundle_creator.sh deploy-appengine dev
     (3) ./deploy/bundle_creator.sh deploy-appengine-v2 dev
     (4) ./deploy/bundle_creator.sh run-docker dev
     (5) ./deploy/bundle_creator.sh request dev
         Request with `cc_emails` and without `cc_emails`.
     (6) Use command to send a v2 CreateBundleRequest with `cc_emails`.
     (7) Manually check result mails' `to` and `cc` fields.
TEST=(1) ./deploy/bundle_creator.sh test-appengine-v2
     (2) ./deploy/bundle_creator.sh test-docker

Change-Id: I95856176cc518c5c17470655a8ba322d25e905c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4360899
Reviewed-by: Yu-An Wang <wyuang@google.com>
Tested-by: Liang-Chieh Chen <jamesqaq@chromium.org>
Commit-Queue: Liang-Chieh Chen <jamesqaq@chromium.org>