tpm_state: Fix matching string
The tpm_state pytest would fail when executed even though the
command output seems to be correct.

This behavior is caused by matching the output string with incorrect
target string. This CL fixes this behavior by removing the space
of the output string and match it with a normal string.

BUG=b:274528785
TEST=make test

Change-Id: Ia19f749902050712917363ce5a61530be336420c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4368966
Tested-by: Louis Chiu <louischiu@google.com>
Commit-Queue: Louis Chiu <louischiu@google.com>
Reviewed-by: Wen-Ting Tseng <wdzeng@chromium.org>