factory: Fix ps_all_pids and switching test list
This CL fixes function ps_all_pids in process_utils which makes the
pytest unable to be stopped in the factory toolkit.

This CL slightly changes some logging behavior in process_utils.

This CL changes the typing of args in Spawn from List to Sequence. This
is to allow user to pass arguments of type other than list, such as a
tuple.

This CL fixes invalid argument in goofy to fix that test list cannot be
switched.

BUG=b:273881230
TEST=make test
TEST=Manually stop a pytest on factory toolkit on DUT by GUI
TEST=Run `factory stop` on DUT
TEST=Switch test list on factory toolkit on DUT

Change-Id: I620a127bcb2468ba03c9c102bc406e7912370459
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4368971
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>