hwid/service: Merge same expected fields in different components
For some categories, there may be multiple expected fields lists
generated, causing one component’s expected fields to possibly partially
overlap with another component’s. In this case, when the result of the
probe is the intersection of the two components’ expected fields, both
two components will be probed. To fix this issue, this CL modifies the
payload generator to check if there is any overlapping between every two
components, and remove the overlapped expected fields from one of the
components if any.

BUG=b:275023407
TEST=./deploy/cros_hwid_service.sh test
TEST=make test

Change-Id: I95268d4369751ad39cd52e78957e2555502f449b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4374281
Reviewed-by: Clark Chung <ckclark@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>