hwid,probe,testlog: use re.fullmatch instead when possible
To prevent the misunderstanding of re.match which only matches prefix,
we should prefer re.fullmatch whenever available.  This CL removes the
trailing '$' and update the usages from {pat,re}.match to
{pat,re}.fullmatch if the string does not ends with a new line[1].

[1] According to the official doc, '$' could match the end of a string
or just before the newline at the end of a string.  For example,
`re.match('xxx$', 'xxx\n')` is a match but
`re.fullmatch('xxx', 'xxx\n')` is not.

BUG=b:278028832
TEST=make test
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I8d47aaa351682f284c7ef3e5467c4957f4247797
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4421520
Commit-Queue: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>