factory: use re.fullmatch instead when possible
To prevent the misunderstanding of re.match which only matches prefix,
we should prefer re.fullmatch whenever available.  This CL removes the
trailing '$' and update the usages from {pat,re}.match to
{pat,re}.fullmatch if the string does not ends with a new line[1].

[1] According to the official doc, '$' could match the end of a string
or just before the newline at the end of a string.  For example,
`re.match('xxx$', 'xxx\n')` is a match but
`re.fullmatch('xxx', 'xxx\n')` is not.

This CL focuses on the folders below:

- gooftool
- hwid_extractor
- test
- umpire

BUG=b:278028832
TEST=make test
TEST=setup/cros_docker.sh umpire test

Change-Id: I5f9496532ca4572013fb66e5a2d953bb1a36c05c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4421521
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>