factory: use re.fullmatch instead when possible
To prevent the misunderstanding of re.match which only matches prefix,
we should prefer re.fullmatch whenever available.  This CL removes the
trailing '$' and update the usages from {pat,re}.match to
{pat,re}.fullmatch if the string does not ends with a new line[1].

[1] According to the official doc, '$' could match the end of a string
or just before the newline at the end of a string.  For example,
`re.match('xxx$', 'xxx\n')` is a match but
`re.fullmatch('xxx', 'xxx\n')` is not.

This CL focuses on the folders below:

- po
- py/device
- py/doc
- py/tools

BUG=b:278028832
TEST=make test

Change-Id: I98c46e2b5d518b550f7652ea7aa2ea8526e0639d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4421522
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>