probe: Add probe for feature management flags
Feature management flags are part of the HWID by design so we need a
probe function for it. The probe is designed in a two-stage fashion.
If RLZ code is set on given DUT, the probe function will get the value
stored in GSC, however if not, it will return values in regard to the
values stored in device data, because we need to create temporary HWID
for pytest defined in b/274033418.

BUG=b:274033431
TEST=make test
run probe eval-function feature_management_flags on DUT

Change-Id: I4567500d7679819c0022c31600dfddd577b19007
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4466728
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>