pytest: Add feature compliance version test
For feature management flow to work, we need to provision
feature compliance version to the GSC INFO page. This test generates
a temporary HWID string and use the probed value to run the
feature compliance check to get the correct value to be provisioned
in later steps.

BUG=b:274033418
TEST=run manual steps on DUT
0. Add custom-made textproto to DUT. Run hwid collect-material and
hwid update-database with the probe defined in b/274033431.
1. run test BrandedChassis defined in b/274032523 to set device data
for chassis_branded.
2. run this pytest and it should pass the test.
3. re-run step 0-2 changing values in textproto and chassis_branded
device data to create all possible combinations.

Change-Id: If034031cec5b23c30fbad630429f22ad844816b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4466729
Tested-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>
Reviewed-by: Stimim Chen <stimim@chromium.org>